% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{betaX}
\alias{betaX}
\title{Simulated exposure effect estimates}
\format{
A data frame with 100 rows and 3 columns:
\describe{
  \item{SNP}{SNP identifier}
  \item{exp1}{Effect on exposure 1}
  \item{exp2}{Effect on exposure 2}
}
}
\usage{
betaX
}
\description{
Simulated exposure effect estimates
}
\keyword{datasets}
