% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwasY}
\alias{gwasY}
\title{Simulated GWAS summary data for outcome Y}
\format{
A data frame with 100 rows and 7 columns:
\describe{
  \item{SNP}{SNP identifier}
  \item{CHR}{Chromosome}
  \item{BP}{Base-pair position}
  \item{A1, A2}{Effect and other alleles}
  \item{BETA}{Estimated effect size}
  \item{SE}{Standard error}
}
}
\usage{
gwasY
}
\description{
Simulated GWAS summary data for outcome Y
}
\keyword{datasets}
